add embeddable bundle
I was able to get my custom widget (built using this cookie-ts-cutter) to be embeddable in an exported HTML file and render on NBviewer after making this change. This was referenced in issue #29 and proposed by @jasongrout (there was also an additional change proposed to the package.json, which I'll include an another commit).